refactor(pr09): rename _types.mts -> types.mts across settings tree (G5h)
git mv 8 files under src/settings/{core,currency,display,gameRules,health,roll,shared/sheet/SettingsTable,skills}/_types.mts to types.mts. Rewrites 27 importer files to drop the leading underscore from the import path. Leaves Foundry-external imports (@common/_types.mjs, @common/abstract/_types.mjs, @client/applications/_types.mjs) untouched.
