+++ src/entities/activeEffects/material/sheet/MaterialStore.mts
@@ -6,8 +6,8 @@ import { getDefaultActiveEffectTabs } from '@effects/BaseActiveEffect/sheet/tabs
 import type { MaterialType } from '@effects/material/Material.mjs';
 import { Material } from '@effects/material/Material.mjs';
 import { PriceData } from '@fields/PriceData.mjs';
-import type { DamageReductionTypesConfig } from '@settings/gameRules/_types.mjs';
 import { GAME_RULES_KEYS } from '@settings/gameRules/constants.mjs';
+import type { DamageReductionTypesConfig } from '@settings/gameRules/types.mjs';
 import { SYSTEM_ID } from '@settings/shared.mjs';
 import type { MultiSelectOption } from '@vc/Fields/FormGroups/types.mjs';
 import type { VueApplicationContext } from '@vueApps/VueAppTypes.mjs';
