+++ src/fields/PriceData.mts
@@ -12,8 +12,8 @@
  * It acts as a snapshot so the GP value is still available if settings change.
  */
 
-import type { CoinageDefinition, CoinStack, CurrencyConfig } from '@settings/currency/_types.mjs';
 import { CURRENCY_KEY, DEFAULT_CURRENCY_CONFIG } from '@settings/currency/constants.mjs';
+import type { CoinageDefinition, CoinStack, CurrencyConfig } from '@settings/currency/types.mjs';
 import { SYSTEM_ID } from '@settings/shared.mjs';
 
 const { DataModel } = foundry.abstract;
