+++ src/fields/PriceField.mts
@@ -15,7 +15,7 @@
  * - Raw CoinStack arrays (from programmatic change building)
  */
 
-import type { CoinStack } from '@settings/currency/_types.mjs';
+import type { CoinStack } from '@settings/currency/types.mjs';
 
 import { PriceData } from './PriceData.mjs';
