+++ src/settings/core/index.mts
@@ -2,6 +2,6 @@
  * Core settings module
  */
 
-export type { FieldChoice, SettingField, SettingsSection } from './_types.mjs';
 export { CORE_KEYS } from './constants.mjs';
 export { registerCoreSettings } from './registration.mjs';
+export type { FieldChoice, SettingField, SettingsSection } from './types.mjs';
