+++ src/settings/currency/sheet/CurrencySettingsApp.vue
@@ -76,9 +76,9 @@
   import type { VueSettingsContext } from '@vueApps/VueSettingsMixin.mjs';
   import { computed, inject, ref, watch } from 'vue';
 
-  import type { CoinageDefinition, CurrencyConfig } from '../_types.mjs';
-  import { coinageVisibilityEveryone } from '../_types.mjs';
   import { CURRENCY_KEY, DEFAULT_CURRENCY_CONFIG, USER_COIN_PREFIX } from '../constants.mjs';
+  import type { CoinageDefinition, CurrencyConfig } from '../types.mjs';
+  import { coinageVisibilityEveryone } from '../types.mjs';
   import CoinageRow from './CoinageRow.vue';
   import { AUTO_ID_PREFIX } from './idFieldUtils.mjs';
