+++ src/settings/gameRules/index.mts
@@ -2,7 +2,6 @@
  * Game Rules settings module
  */
 
-export type { DamageReductionTypeEntry, DamageReductionTypesConfig, DiagonalMovementRule, ExperienceRate } from './_types.mjs';
 export {
   DEFAULT_DAMAGE_REDUCTION_TYPES,
   DIAGONAL_MOVEMENT_CHOICES,
@@ -12,3 +11,4 @@ export {
 } from './constants.mjs';
 export { registerGameRulesSettings } from './registration.mjs';
 export { GameRulesSettingsConfig } from './sheet/index.mjs';
+export type { DamageReductionTypeEntry, DamageReductionTypesConfig, DiagonalMovementRule, ExperienceRate } from './types.mjs';
