+++ src/settings/gameRules/types.mts
