+++ src/settings/health/index.mts
@@ -2,7 +2,6 @@
  * Health settings module
  */
 
-export type { HealthConfig, HitDieConfig } from './_types.mjs';
 export {
   DEFAULT_HEALTH_CONFIG,
   HEALTH_KEY,
@@ -10,3 +9,4 @@ export {
 } from './constants.mjs';
 export { registerHealthSettings } from './registration.mjs';
 export { HealthSettingsConfig } from './sheet/index.mjs';
+export type { HealthConfig, HitDieConfig } from './types.mjs';
