+++ src/settings/shared/sheet/SettingsTable/index.mts
@@ -2,7 +2,7 @@
  * SettingsTable component exports
  */
 
-export type { ColumnRenderContext, IdConfig, SettingsTableColumn, SettingsTableItem } from './_types.mjs';
-export { AUTO_ID_MARKER, DEFAULT_ID_PREFIX } from './_types.mjs';
 export { default as SettingsTable } from './SettingsTable.vue';
 export { default as SettingsTableRow } from './SettingsTableRow.vue';
+export type { ColumnRenderContext, IdConfig, SettingsTableColumn, SettingsTableItem } from './types.mjs';
+export { AUTO_ID_MARKER, DEFAULT_ID_PREFIX } from './types.mjs';
