+++ src/settings/skills/index.mts
@@ -2,7 +2,6 @@
  * Skills settings module
  */
 
-export type { CustomSkill, SkillSettings } from './_types.mjs';
 export {
   DEFAULT_SKILL_SETTINGS,
   SKILLS_KEY,
@@ -10,3 +9,4 @@ export {
 } from './constants.mjs';
 export { registerSkillsSettings } from './registration.mjs';
 export { SkillSettingsConfig } from './sheet/index.mjs';
+export type { CustomSkill, SkillSettings } from './types.mjs';
